+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/commands/Command.kt
@@ -1,14 +1,15 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.utils.commands
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.checkedMessagesFlow
-import com.github.insanusmokrassar.AutoPostTelegramBot.messagesListener
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.flow.collectWithErrors
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageEntity.BotCommandMessageEntity
 import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.CommonMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.TextContent
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.UpdateReceiver
+import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.mapNotNull
 
@@ -18,8 +19,10 @@ abstract class Command {
     protected abstract val commandRegex: Regex
 
     init {
-        messagesListener.subscribe {
-            invoke(it)
+        CoroutineScope(Dispatchers.Default).launch {
+            checkedMessagesFlow.collectWithErrors {
+                invoke(it)
+            }
         }
     }